--- music/demos/soundfont.html
@@ -1,51 +1,0 @@
-<!-- Copyright 2019 Google Inc. All Rights Reserved.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-==============================================================================-->
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
-  <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1, user-scalable=yes">
-  <link rel="stylesheet" href="./style.css">
-  <title>mm.SoundFontPlayer</title>
-  <style>
-    td {
-      min-width: 100px;
-    }
-  </style>
-</head>
-
-<body>
-  <h1>mm.SoundFontPlayer</h1>
-  <p>There are several SoundFonts that you can use with the <code>mm.SoundFontPlayer</code>,
-    for more realistic sounding instruments.
-
-  <h2>Piano: <code>salamander</code></h2>
-  <section>
-    <p>Audio samples from <a href="https://archive.org/details/SalamanderGrandPianoV3">Salamander Grand Piano</a>.</p>
-    <button id="btnSalamander">Play</button>
-  </section>
-
-  <h2>Multi-instrument: <code>sgm_plus</code></h2>
-  <section>
-    <p>Audio samples based on <a href="https://www.polyphone-soundfonts.com/en/files/27-instrument-sets/256-sgm-v2-01">SGM</a> with modifications by <a href="https://sites.google.com/site/soundfonts4u/" rel="nofollow">John Nebauer</a></p>
-    <label>Instrument: <select id="select" class="button"></select></label>
-    <button id="btnSgmPlus">Play</button>
-  </section>
-
-  <h2>Percussion: <code>jazz_kit</code></h2>
-  <section>
-    <p>Audio samples from <a href="https://musical-artifacts.com/artifacts/686">Jazz Kit (EXS)</a> by Lithalean</p>
-    <button id="btnJazzKit">Play</button>
-  </section>
-  <script src="soundfont_bundle.js"></script>
-</body>
-</html>
